fix(capture): 修复 HDR处理时间日志输出的精度问题
- 将 int64 类型的处理时间转换为 int 类型，以避免精度丢失导致的负值输出
- 优化日志输出格式，确保正确记录 HDR 处理时间

+++ app/src/main/cpp/capture.cpp
@@ -76,7 +76,9 @@ void CaptureManager::process(int jobId) {
         // cv::cvtColor(yuvMat, rgbMat, cv::COLOR_YUV420sp2RGB);
         merge_mertens->process(jobs[jobId]->frames, fusion);
         auto hdr_t= cv::getTickCount();
-        LOGD("%s, hdr processing cost %d s", __FUNCTION__, (hdr_t - start_t) / cv::getTickFrequency());
+        // int64 必须要转成 int，否则输出会丢失精度后变成负值
+        LOGD("%s, hdr processing cost %d s", __FUNCTION__, static_cast<int>((hdr_t - start_t) /
+        cv::getTickFrequency()));
 
         // // default set jpeg quality to 95
         // std::vector<int> params{cv::IMWRITE_JPEG_QUALITY, 95};
